renamed module world_timer.py to improve readability
